\name{dlcdfALD}
\alias{dlcdfALD}
\title{Evaluate the CDF of an Asymmetric Laplace Distribution}
\description{
  \code{dlcdfALD} returns a CDF evaluation from an Asymmetric Laplace Distribution.
}
\usage{
dlcdfALD( 
  x, 
  tau, 
  mu, 
  sigma )
}
\arguments{
  \item{x}{Point to evaluate.}
  
  \item{tau}{The ALD weighting parameter.}
  
  \item{mu}{The split point for weighted tails.}
  
  \item{sigma}{The spread parameter. }
  
}

\author{Daniel Dempsey and Jason Wyse}
